Optimised

+++ backend/controllers/task.js
@@ -28,18 +28,23 @@ async function provideTask(req, res) {
 }
 
 async function deleteTask(req, res) {
+  if(!req.user._id) return res.json({message:"not found"})
+
   const id = req.params.id;
+  const userId = req.user._id;
   try {
     const deleteData = await taskModel.findByIdAndDelete(id);
+    const userData = await taskModel.find({userId: userId});
     if (!deleteData) return res.status(400).json({ msg: "No such task" });
-    else return res.json({ msg: "success", deleted: deleteData });
+    else return res.json({ msg: "success", deleted: deleteData, userData: userData });
   } catch (err) {
     return res.json({ Error: err });
   }
 }
 
 async function updateTask(req, res) {
-  console.log("Update");
+  if(!req.user._id) return res.json({message:"not found"})
+    const userId = req.user._id;
   const id = req.params.id;
   const taskData = req.body;
   try {
@@ -48,7 +53,7 @@ async function updateTask(req, res) {
       { taskInfo: taskData.taskInfo },
       { new: true }
     );
-    const updatedTaskData = await taskModel.find();
+    const updatedTaskData = await taskModel.find({userId: userId});
     return res.json(updatedTaskData);
   } catch (error) {
     return res.status(400).json({ msg: "Error in updating task" });
@@ -58,14 +63,15 @@ async function updateTask(req, res) {
 async function updateToggleCheck(req, res) {
   const id = req.params.id;
   const taskData = req.body;
-
+  if(!req.user._id) return res.json({message:"not found"})
+  const userId = req.user._id;
   try {
     const updateToggle = await taskModel.findByIdAndUpdate(
       id,
       { toggleCheck: !taskData.toggleCheck },
       { new: true }
     );
-    const updatedTaskData = await taskModel.find();
+    const updatedTaskData = await taskModel.find({userId: userId});
     return res.json(updatedTaskData);
   } catch (error) {
     return res.status(400).json({ msg: "Error in updating toggle" });
@@ -73,14 +79,15 @@ async function updateToggleCheck(req, res) {
 }
 
 async function updateStarMark(req, res) {
+  if(!req.user._id) return res.json({message:"not found"})
   const id = req.params.id;
   const taskData = req.body;
-
+  const userId = req.user._id;
   try {
     const updateStarMark = await taskModel.findByIdAndUpdate(id, {
       starMarked: !taskData.starMarked,
     });
-    const updatedStarData = await taskModel.find();
+    const updatedStarData = await taskModel.find({userId: userId});
     return res.json(updatedStarData);
   } catch (err) {
     return res.status(400).json({ msg: "Error in updating star" });
